requirements.txt
flask
google-generativeai
pillow
reportlab
qrcode[pil]
python-dotenv
requests
werkzeug
